Extract shared AbstractPostgresPersistenceAdapter for core
Same pattern as the digest/journal extractions: the ~40 pure-SQL core
persistence methods (instances, charts, ongoing activities, deferred
events, external ids) are written once against $N positional
placeholders in an abstract base that stays generic over ConnectionType,
so core-pg and core-pglite keep their own connection type and
withTransaction signature unchanged. Each driver now only implements
runQuery plus its connection/transaction/death-note-polling layer.

State bytea decoding is unified via decodeBytea in the base instead of
being duplicated per driver. Removes pg-bind from core-pg now that all
queries use positional parameters.
