+++ packages/core-persistence/src/index.ts
@@ -1,2 +1,3 @@
+export * from './AbstractPostgresPersistenceAdapter';
 export * from './EntryTypes';
 export * from './PersistenceAdapter';
