+++ packages/core-pg/src/PostgresPersistenceAdapter.ts
@@ -1,60 +1,18 @@
 import path from 'node:path';
-import {
-  type PersistedChart,
-  type PersistedDeferredEvent,
-  PersistenceAdapter,
-} from '@telia-oss/xjog-core-persistence';
-import {
-  ChartIdentifier,
-  type ChartReference,
-  getCorrelationIdentifier,
-} from '@telia-oss/xjog-util';
+import { AbstractPostgresPersistenceAdapter } from '@telia-oss/xjog-core-persistence';
 import migrationRunner from 'node-pg-migrate';
-import { type ClientConfig, Pool, type PoolClient, type PoolConfig } from 'pg';
-import bind from 'pg-bind';
-
-import type {
-  EventObject,
-  State,
-  StateConfig,
-  StateSchema,
-  Typestate,
-} from 'xstate';
+import {
+  type ClientConfig,
+  Pool,
+  type PoolClient,
+  type PoolConfig,
+  type QueryResultRow,
+} from 'pg';
 
 /**
  * Options for instantiating {@link PostgresPersistenceAdapter}.
  */
 export type PostgreSQLPersistenceAdapterOptions = Record<string, never>;
-
-/**
- * Chart row directly from the SQL query
- */
-type PostgreSQLChartRow = {
-  timestamp: number;
-  ownerId: string;
-  machineId: string;
-  chartId: string;
-  parentMachineId: string;
-  parentChartId: string;
-  state: Buffer;
-  paused: boolean;
-};
-
-/**
- * Deferred event row directly from the SQL query
- */
-type PostgreSQLDeferredEventRow = {
-  id: number;
-  machineId: string;
-  chartId: string;
-  eventId: string;
-  eventTo: string;
-  event: string;
-  timestamp: number;
-  delay: number;
-  due: number;
-  lock: string;
-};
 
 /**
  * Use the static method [connect]{@link PostgresPersistenceAdapter.connect} to
@@ -65,7 +23,9 @@ type PostgreSQLDeferredEventRow = {
  * @extends PersistenceAdapter
  * @hideconstructor
  */
-export class PostgresPersistenceAdapter extends PersistenceAdapter<PoolClient> {
+export class PostgresPersistenceAdapter extends AbstractPostgresPersistenceAdapter<
+  Pool | PoolClient
+> {
   public readonly component = 'persistence';
   public readonly type = 'pg';
 
@@ -147,279 +107,13 @@ export class PostgresPersistenceAdapter extends PersistenceAdapter<PoolClient> {
     }
   }
 
-  public async countAliveInstances(
+  protected async runQuery<T>(
     connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'SELECT COUNT(*) AS "instanceCount" FROM "instances" WHERE "dying"=FALSE',
-    );
-    return Number(result.rows[0].instanceCount);
-  }
-
-  protected async insertInstance(
-    id: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      bind('INSERT INTO "instances" ("instanceId") VALUES (:id)', { id }),
-    );
-  }
-
-  protected async markInstanceDying(
-    id: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    // Refresh "timestamp" so it marks when the row entered the dying state.
-    // reapDeadInstances ages rows out relative to this, so a long-lived
-    // instance is not reaped immediately after a graceful shutdown.
-    await connection.query(
-      bind(
-        'UPDATE "instances" SET "dying"=TRUE, "timestamp"=now() ' +
-          'WHERE "instanceId"=:id',
-        { id },
-      ),
-    );
-  }
-
-  protected async reapDeadInstances(
-    retentionMs: number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      bind(
-        'DELETE FROM "instances" ' +
-          'WHERE "dying"=TRUE ' +
-          'AND "timestamp" < now() - make_interval(secs => :seconds)',
-        { seconds: retentionMs / 1000 },
-      ),
-    );
-  }
-
-  protected async markAllInstancesDying(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    // Only flip rows that are currently alive, stamping when they became
-    // dying. Rows already dying keep their original timestamp so they can age
-    // out via reapDeadInstances instead of being perpetually refreshed.
-    await connection.query(
-      'UPDATE "instances" SET "dying"=TRUE, "timestamp"=now() ' +
-        'WHERE "dying"=FALSE',
-    );
-  }
-
-  protected async updateInstanceHeartbeat(
-    id: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    // For alive rows "timestamp" doubles as the liveness heartbeat; for dying
-    // rows it is the time of death and must not be refreshed.
-    await connection.query(
-      bind(
-        'UPDATE "instances" SET "timestamp"=now() ' +
-          'WHERE "instanceId"=:id AND "dying"=FALSE',
-        { id },
-      ),
-    );
-  }
-
-  protected async markStaleInstancesAsDying(
-    id: string,
-    stalenessMs: number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'UPDATE "instances" SET "dying"=TRUE, "timestamp"=now() ' +
-          'WHERE "dying"=FALSE AND "instanceId" <> :id ' +
-          'AND "timestamp" < now() - make_interval(secs => :seconds)',
-        { id, seconds: stalenessMs / 1000 },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  protected async pauseChartsWithoutLiveOwner(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'UPDATE "charts" SET "paused"=TRUE ' +
-        'WHERE "paused"=FALSE AND ("ownerId" IS NULL OR "ownerId" NOT IN (' +
-        '  SELECT "instanceId" FROM "instances" WHERE "dying"=FALSE' +
-        '))',
-    );
-    return result.rowCount ?? 0;
-  }
-
-  protected async pauseChartsOwnedBy(
-    id: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'UPDATE "charts" SET "paused"=TRUE ' +
-          'WHERE "paused"=FALSE AND "ownerId"=:id',
-        { id },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  protected async releaseDeferredEventsWithoutLiveOwner(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'UPDATE "deferredEvents" SET "lock"=NULL ' +
-        'WHERE "lock" IS NOT NULL AND "lock" NOT IN (' +
-        '  SELECT "instanceId" FROM "instances" WHERE "dying"=FALSE' +
-        ')',
-    );
-    return result.rowCount ?? 0;
-  }
-
-  protected async claimPausedChart(
-    instanceId: string,
-    ref: ChartReference,
-    requireIdle: boolean,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<boolean> {
-    const result = await connection.query(
-      bind(
-        'UPDATE "charts" SET "ownerId"=:instanceId, "paused"=FALSE ' +
-          'WHERE "machineId"=:machineId AND "chartId"=:chartId ' +
-          'AND "paused"=TRUE ' +
-          (requireIdle
-            ? 'AND NOT EXISTS (' +
-              '  SELECT 1 FROM "ongoingActivities" ' +
-              '  WHERE "machineId"=:machineId AND "chartId"=:chartId' +
-              ') '
-            : ''),
-        {
-          instanceId,
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-        },
-      ),
-    );
-    return (result.rowCount ?? 0) > 0;
-  }
-
-  protected async deleteOngoingActivitiesForChart(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'DELETE FROM "ongoingActivities" ' +
-          'WHERE "machineId"=:machineId AND "chartId"=:chartId',
-        { machineId: ref.machineId, chartId: ref.chartId },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  protected async markAllChartsPaused(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query('UPDATE "charts" SET "paused"=TRUE');
-  }
-
-  protected async countPausedCharts(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'SELECT COUNT(*) AS "chartCount" FROM "charts" WHERE "paused"=TRUE',
-    );
-    return result.rows[0].chartCount;
-  }
-
-  protected async getPausedChartIds(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<ChartReference[]> {
-    const result = await connection.query(
-      'SELECT "machineId", "chartId" FROM "charts" WHERE "paused"=TRUE',
-    );
-
-    return result.rows;
-  }
-
-  protected async getPausedChartWithNoOngoingActivitiesIds(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<ChartReference[]> {
-    const result = await connection.query(
-      'SELECT "machineId", "chartId" FROM "charts" ' +
-        'WHERE "paused" = TRUE AND NOT EXISTS (' +
-        '  SELECT * FROM "ongoingActivities" ' +
-        '  WHERE "machineId" = "charts"."machineId"' +
-        '    AND "chartId" = "charts"."chartId"' +
-        ')',
-    );
-
-    return result.rows;
-  }
-
-  public async countOwnCharts(
-    instanceId: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'SELECT COUNT(*) AS "chartCount" FROM "charts" ' +
-          'WHERE "ownerId" = :instanceId',
-        { instanceId },
-      ),
-    );
-    return Number(result.rows[0].chartCount);
-  }
-
-  protected async deleteOngoingActivitiesForPausedCharts(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'DELETE FROM "ongoingActivities" WHERE NOT EXISTS (' +
-        '  SELECT * FROM "charts" WHERE ' +
-        '    "machineId" = "ongoingActivities"."machineId" AND' +
-        '    "chartId" = "ongoingActivities"."chartId" AND' +
-        '    "paused" = TRUE' +
-        ')',
-    );
-
-    return result.rowCount ?? 0;
-  }
-
-  protected async changeOwnerAndResumePausedCharts(
-    id: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      bind(
-        'UPDATE "charts" ' +
-          'SET "paused" = FALSE, "ownerId" = :id ' +
-          'WHERE "paused" = TRUE',
-        { id },
-      ),
-    );
-  }
-
-  // TODO use row id:s instead, can use IN, faster; also do in batches
-  protected async changeOwnerAndResumeCharts(
-    instanceId: string,
-    refs: ChartReference[],
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    for (const ref of refs) {
-      await connection.query(
-        bind(
-          'UPDATE "charts" ' +
-            'SET "paused" = FALSE, "ownerId" = :instanceId ' +
-            'WHERE "machineId" = :machineId AND "chartId" = :chartId ',
-          {
-            instanceId,
-            machineId: ref.machineId,
-            chartId: ref.chartId,
-          },
-        ),
-      );
-    }
+    sql: string,
+    params?: unknown[],
+  ): Promise<{ rows: T[]; rowCount: number }> {
+    const result = await connection.query<QueryResultRow>(sql, params);
+    return { rows: result.rows as T[], rowCount: result.rowCount ?? 0 };
   }
 
   // TODO use listening instead, slicker
@@ -439,12 +133,10 @@ export class PostgresPersistenceAdapter extends PersistenceAdapter<PoolClient> {
         return;
       }
 
-      const result = await connection.query(
-        bind(
-          'SELECT COUNT(*) as "dying" FROM "instances" ' +
-            'WHERE "instanceId" = :instanceId AND "dying" = TRUE',
-          { instanceId },
-        ),
+      const result = await connection.query<{ dying: number }>(
+        'SELECT COUNT(*) as "dying" FROM "instances" ' +
+          'WHERE "instanceId" = $1 AND "dying" = TRUE',
+        [instanceId],
       );
 
       // Re-check after the await: cancel() may have run meanwhile, or a
@@ -470,495 +162,6 @@ export class PostgresPersistenceAdapter extends PersistenceAdapter<PoolClient> {
       if (timer) {
         clearInterval(timer);
       }
-    };
-  }
-
-  protected async insertChart<
-    TContext,
-    TEvent extends EventObject,
-    TStateSchema extends State<any>,
-    TTypeState extends Typestate<any>,
-  >(
-    instanceId: string,
-    ref: ChartReference,
-    parentRef: ChartReference | null,
-    state: State<TContext, TEvent, TStateSchema, TTypeState>,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      bind(
-        'INSERT INTO charts ' +
-          '(' +
-          '  "ownerId", "machineId", "chartId", ' +
-          '  "parentMachineId", "parentChartId", "state"' +
-          ') ' +
-          'VALUES (' +
-          '  :instanceId, :machineId, :chartId, ' +
-          '  :parentMachineId, :parentChartId, :state ' +
-          ')',
-        {
-          instanceId,
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-          parentMachineId: parentRef?.machineId ?? null,
-          parentChartId: parentRef?.chartId ?? null,
-          state: Buffer.from(JSON.stringify(state)),
-        },
-      ),
-    );
-  }
-
-  protected async chartExists(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<boolean> {
-    const result = await connection.query(
-      bind(
-        'SELECT 1 FROM "charts" ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId',
-        {
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-        },
-      ),
-    );
-
-    return result.rows.length > 0;
-  }
-
-  protected async readChart<TContext, TEvent extends EventObject>(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<PersistedChart<TContext, TEvent> | null> {
-    const result = await connection.query<PostgreSQLChartRow>(
-      bind(
-        'SELECT * FROM "charts" ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId ',
-        { machineId: ref.machineId, chartId: ref.chartId },
-      ),
-    );
-
-    if (!result.rowCount) {
-      return null;
-    }
-
-    return PostgresPersistenceAdapter.parseSqlChartRow<TContext, TEvent>(
-      result.rows[0],
-    );
-  }
-
-  protected async updateChartState<TContext, TEvent extends EventObject>(
-    ref: ChartReference,
-    state: State<TContext, TEvent, StateSchema<TContext>, Typestate<TContext>>,
-    expectedOwnerId: string | null = null,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'UPDATE "charts" SET "state" = :state ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId' +
-          (expectedOwnerId !== null ? ' AND "ownerId" = :expectedOwnerId' : ''),
-        {
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-          state: Buffer.from(JSON.stringify(state)),
-          ...(expectedOwnerId !== null ? { expectedOwnerId } : {}),
-        },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  /**
-   * @returns Number of deleted records
-   */
-  protected async deleteChart(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'DELETE FROM "charts" WHERE "machineId" = :machineId AND "chartId" = :chartId',
-        {
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-        },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  public async isActivityRegistered(
-    ref: ChartReference,
-    activityId: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<boolean> {
-    const result = await connection.query(
-      bind(
-        'SELECT 1 FROM "ongoingActivities" ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId AND "activityId" = :activityId',
-        {
-          machineId: ref.machineId,
-          chartId: ref.chartId,
-          activityId,
-        },
-      ),
-    );
-
-    return result.rows.length > 0;
-  }
-
-  public async registerActivity(
-    ref: ChartReference,
-    activityId: string,
-    cid: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    const response = await connection.query(
-      bind(
-        'INSERT INTO "ongoingActivities" ' +
-          '  ("machineId", "chartId", "activityId") ' +
-          'VALUES (:machineId, :chartId, :activityId) ' +
-          'ON CONFLICT ("machineId", "chartId", "activityId") ' +
-          '  DO NOTHING ',
-        { machineId: ref.machineId, chartId: ref.chartId, activityId },
-      ),
-    );
-  }
-
-  public async unregisterActivity(
-    ref: ChartReference,
-    activityId: string,
-    cid: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    const response = await connection.query(
-      bind(
-        'DELETE FROM "ongoingActivities" ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId AND "activityId" = :activityId',
-        { machineId: ref.machineId, chartId: ref.chartId, activityId },
-      ),
-    );
-  }
-
-  /**
-   * @returns Number of deleted records
-   */
-  protected async deleteAllDeferredEvents(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      bind(
-        'DELETE FROM "deferredEvents" WHERE "machineId" = :machineId AND "chartId" = :chartId',
-        { machineId: ref.machineId, chartId: ref.chartId },
-      ),
-    );
-    return result.rowCount ?? 0;
-  }
-
-  public async getExternalIdentifiers(
-    key: string,
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<string[]> {
-    const result = await connection.query<{ value: string }>(
-      bind(
-        'SELECT "value" FROM "externalId" ' +
-          'WHERE "machineId" = :machineId AND "chartId" = :chartId AND "key" = :key',
-        { machineId: ref.machineId, chartId: ref.chartId, key },
-      ),
-    );
-    return result.rows.map((row) => row.value);
-  }
-
-  public async getChartByExternalIdentifier(
-    key: string,
-    value: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<ChartReference | null> {
-    const result = await connection.query<ChartReference>(
-      'SELECT "machineId", "chartId" FROM "externalId" ' +
-        'WHERE "key"=$1 AND "value"=$2 FOR SHARE',
-      [key, value],
-    );
-    return result.rows[0] ?? null;
-  }
-
-  public async registerExternalId(
-    ref: ChartReference,
-    key: string,
-    value: string,
-    cid: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      'INSERT INTO "externalId" ' +
-        '  ("machineId", "chartId", "key", "value") ' +
-        'VALUES ' +
-        '  ($1, $2, $3, $4) ' +
-        'ON CONFLICT ( ' +
-        '  "key", "value" ' +
-        ') DO UPDATE SET ' +
-        '  "machineId" = $1, "chartId" = $2 ',
-      [ref.machineId, ref.chartId, key, value],
-    );
-  }
-
-  public async dropExternalId(
-    key: string,
-    value: string,
-    cid: string,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'DELETE FROM "externalId" WHERE "key"=$1 AND "value"=$2',
-      [key, value],
-    );
-    return result.rowCount ?? 0;
-  }
-
-  /**
-   * @returns Number of deleted records
-   */
-  protected async deleteExternalIdentifiers(
-    ref: ChartReference,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'DELETE FROM "externalId" WHERE "machineId"=$1 AND "chartId"=$2',
-      [ref.machineId, ref.chartId],
-    );
-    return result.rowCount ?? 0;
-  }
-
-  /** Corresponds to {@link PostgreSQLDeferredEventRow} */
-  private readonly deferredEventSelectFields =
-    '  "id", "machineId", "chartId", "lock", ' +
-    '  "eventId", "eventTo", "event", "delay", ' +
-    '  extract(epoch from "timestamp") * 1000 as "timestamp", ' +
-    '  extract(epoch from "due") * 1000 as "due" ';
-
-  protected async readDeferredEventRow(
-    id: number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<PersistedDeferredEvent | null> {
-    const result = await connection.query<PostgreSQLDeferredEventRow>(
-      'SELECT ' +
-        this.deferredEventSelectFields +
-        'FROM "deferredEvents" ' +
-        'WHERE "id"=$1 ' +
-        'FOR SHARE',
-      [id],
-    );
-
-    if (!result.rowCount) {
-      return null;
-    }
-
-    return PostgresPersistenceAdapter.parseSqlDeferredEventRow(result.rows[0]);
-  }
-
-  protected async readDeferredEventByEventId(
-    ref: ChartReference,
-    eventId: string | number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<PersistedDeferredEvent | null> {
-    const result = await connection.query<PostgreSQLDeferredEventRow>(
-      'SELECT ' +
-        this.deferredEventSelectFields +
-        'FROM "deferredEvents" ' +
-        'WHERE "machineId"=$1 AND "chartId"=$2 AND "eventId"=$3 ' +
-        'FOR SHARE',
-      [ref.machineId, ref.chartId, JSON.stringify(eventId)],
-    );
-
-    if (!result.rowCount) {
-      return null;
-    }
-
-    return PostgresPersistenceAdapter.parseSqlDeferredEventRow(result.rows[0]);
-  }
-
-  /**
-   * Read a batch of deferred events and mark them taken
-   * @param instanceId
-   * @param lookAhead
-   * @param batchSize
-   * @param connection
-   * @protected
-   */
-  protected async readDeferredEventRowBatch(
-    instanceId: string,
-    lookAhead: number,
-    batchSize: number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<PersistedDeferredEvent[]> {
-    const result = await connection.query(
-      'WITH updated AS ( ' +
-        '  UPDATE "deferredEvents" ' +
-        '  SET "lock"=$3 ' +
-        '  WHERE "id" IN ' +
-        '  (' +
-        '    SELECT "id" FROM "deferredEvents" ' +
-        '    WHERE "due"<(transaction_timestamp() + make_interval(secs => $1::bigint / 1000)) ' +
-        '      AND "lock" IS NULL ' +
-        '    ORDER BY "due" ASC, "id" ASC ' +
-        '    LIMIT $2::bigint' +
-        '  ) ' +
-        '  RETURNING * ' +
-        ') ' +
-        'SELECT ' +
-        this.deferredEventSelectFields +
-        'FROM updated ' +
-        'ORDER BY "due" ASC, "id" ASC',
-      [lookAhead, batchSize, instanceId],
-    );
-
-    return result.rows.map(PostgresPersistenceAdapter.parseSqlDeferredEventRow);
-  }
-
-  public async releaseDeferredEvent(
-    ref: ChartReference,
-    eventId: string | number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query(
-      'UPDATE "deferredEvents" SET "lock"=NULL ' +
-        'WHERE "machineId"=$1 AND "chartId"=$2 AND "eventId"=$3',
-      [ref.machineId, ref.chartId, JSON.stringify(eventId)],
-    );
-  }
-
-  protected async unmarkAllDeferredEventsForProcessing(
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<void> {
-    await connection.query('UPDATE "deferredEvents" SET "lock"=NULL');
-  }
-
-  protected async insertDeferredEvent(
-    PersistedDeferredEvent: Omit<PersistedDeferredEvent, 'id'>,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<PersistedDeferredEvent | null> {
-    const timestamp = PersistedDeferredEvent.timestamp ?? Date.now();
-    const due =
-      PersistedDeferredEvent.due ??
-      timestamp + Math.ceil(PersistedDeferredEvent.delay);
-
-    let toFieldStringRepresentation = null;
-
-    if (PersistedDeferredEvent.eventTo) {
-      if (
-        typeof PersistedDeferredEvent.eventTo === 'string' ||
-        typeof PersistedDeferredEvent.eventTo === 'number'
-      ) {
-        toFieldStringRepresentation = PersistedDeferredEvent.eventTo;
-      } else {
-        const chartIdentifierFromToField = ChartIdentifier.from(
-          PersistedDeferredEvent.eventTo,
-        );
-        if (chartIdentifierFromToField) {
-          toFieldStringRepresentation =
-            chartIdentifierFromToField.uri.toString();
-        } else if (
-          typeof PersistedDeferredEvent.eventTo === 'object' &&
-          'id' in PersistedDeferredEvent.eventTo &&
-          PersistedDeferredEvent.eventTo.id
-        ) {
-          toFieldStringRepresentation = PersistedDeferredEvent.eventTo.id;
-        }
-      }
-    }
-
-    const result = await connection.query<PostgreSQLDeferredEventRow>(
-      bind(
-        'INSERT INTO "deferredEvents" (' +
-          '  "machineId", "chartId", ' +
-          '  "eventId", "eventTo", "event", ' +
-          '  "timestamp", "delay", ' +
-          '  "due" ' +
-          ') VALUES (' +
-          '  :machineId, :chartId, ' +
-          '  :eventId, :eventTo, :event, ' +
-          '  :timestamp_iso::timestamptz, :delay::bigint, ' +
-          '  :due_iso::timestamptz' +
-          ') ' +
-          'RETURNING ' +
-          this.deferredEventSelectFields,
-        {
-          machineId: PersistedDeferredEvent.ref.machineId,
-          chartId: PersistedDeferredEvent.ref.chartId,
-
-          eventId: JSON.stringify(PersistedDeferredEvent.eventId),
-          eventTo: JSON.stringify(toFieldStringRepresentation),
-          event: JSON.stringify(PersistedDeferredEvent.event),
-
-          timestamp_iso: new Date(timestamp).toISOString(),
-          delay: Math.ceil(PersistedDeferredEvent.delay),
-          due_iso: new Date(due).toISOString(),
-        },
-      ),
-    );
-
-    if (!result.rowCount) {
-      return null;
-    }
-
-    return PostgresPersistenceAdapter.parseSqlDeferredEventRow(result.rows[0]);
-  }
-
-  protected async deleteDeferredEvent(
-    id: number,
-    connection: Pool | PoolClient = this.pool,
-  ): Promise<number> {
-    const result = await connection.query(
-      'DELETE FROM "deferredEvents" WHERE "id"=$1 ',
-      [id],
-    );
-
-    return result.rowCount ?? 0;
-  }
-
-  private static parseSqlChartRow<TContext, TEvent extends EventObject>(
-    row: PostgreSQLChartRow,
-  ): PersistedChart<TContext, TEvent> {
-    return {
-      timestamp: Number(row.timestamp),
-
-      ownerId: row.ownerId,
-      ref: {
-        machineId: row.machineId,
-        chartId: row.chartId,
-      },
-      parentRef: row.parentChartId
-        ? {
-            machineId: row.parentMachineId,
-            chartId: row.parentChartId,
-          }
-        : null,
-
-      state: JSON.parse(row.state.toString()) as StateConfig<TContext, TEvent>,
-
-      paused: row.paused,
-    };
-  }
-
-  private static parseSqlDeferredEventRow(
-    row: PostgreSQLDeferredEventRow,
-  ): PersistedDeferredEvent {
-    return {
-      id: row.id,
-      timestamp: Number(row.timestamp),
-      ref: {
-        machineId: row.machineId,
-        chartId: row.chartId,
-      },
-      eventId: JSON.parse(row.eventId),
-      eventTo: JSON.parse(row.eventTo),
-      event: JSON.parse(row.event),
-      delay: row.delay,
-      due: row.due,
-      lock: row.lock,
     };
   }
 }
