--- packages/core-pg/src/pg-bind.d.ts
@@ -1,1 +1,0 @@
-declare module 'pg-bind';
